Make real mazes with recursive backtrace
Filled out the C++ map generator to use recursive backtracking to create a real maze.
This involved noticing an error in the 2D array code, where it was returning [] by value not reference, meaning you couldn't modify the contents of the array.
Also messed up the x and y in the MapGenerator transform logic and had to flip them.