Dungeon styling
Styled up the maze using the MedievalDungeon pack - selectively migrating assets as I used them. Included getting the lighting to come from this, an ambient dungeon feel.
Functionally, followed through on what I wanted with map tiles in blueprint, which is the ability to swap out varients at runtime. Passages between tiles can be empty, or can be a door. Walls can be stone, or decorated with arches and a torch.