Add attack animation, trigger on mouse click
Added a swinging attack animation (very bad) that is triggered when you attack - currently left mouse click or controller left.
Handled this in a new BP PlayerController, so it works the same for the minotaur.
Also added 'backspace' as another alias for the character swap, since alt-tab swaps active window for me.
Brought in an axe mesh/skeleton to attach to the hand, haven't done the attaching yet.