Add lootable weapons you can equip, improve swing animation
Made a slightly better swing animation, and hooked a socket up on the player to attach the axe to. You get the axe via a lootable weapon blueprint, which is set to appear in map tiles with a 25% chance. Played around with some VFX for that.
Weapons are set up in a data driven way that should make adding more weapons easy. Damage is stubbed in but not implemented.