Add 2nd weapon type, attack anim swapping
Move combat stuff into a component, in the process of making it properly able to swap weapons. BUG - loot weapon meshes don't always match the actual weapon type?? Can't see the reason right now. When it does match things work as expected - if the weapon you're picking up is different than your current one, it equips the new one. This includes changing the attack animations - and there is a default punch animation for now weapon. Which is what the Minotaur uses, since it isn't set up to equip weapons.