Very basic combat (weapon collides and deals damage)
Weapons are now blueprints which lets them tweak a collision volume to match the killy end. Health component exists and can be used to make an actor take damage. When the player attacks, collision from their weapon will result in damage being dealt equal to the weapon damage in the weapon data table (and punching defaults to 1).
The flow is sorta there now, but the details are bad. Being able to restart attacks but that not clearing the current attack is one example.