Refactor how map tiles are generated to be replication-friendly
Key issue with the previous implementation was that trying to do a lot of creating and immediately deleting or changing child actors at construction time was a bad idea, it inconsistently replicated to clients.
The paradigm now is to create the desired actor from the start. So instead of making 4 dungeon walls and then going through to delete some or change them into walls, it only creates a wall if we want a wall, and if we want a passage it leaves it blank or makes a door.